+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 - build: replace sprintf by snprintf due to upgraded MacOS compiler, change linking of googletest [PR #1362]
 - storage daemon: fix crash on volume swap [PR #1360]
 - systemtests: `rename virtualfull` -> `virtualfull-basic` ,`bareos` -> `bareos-basic`, `bconsole` -> `bconsole-basic` [PR #1363]
+- webui: enable sorting on version column [PR #1368]
 
 ## [22.0.1] - 2023-01-02
 
@@ -441,4 +442,5 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 [PR #1360]: https://github.com/bareos/bareos/pull/1360
 [PR #1362]: https://github.com/bareos/bareos/pull/1362
 [PR #1363]: https://github.com/bareos/bareos/pull/1363
+[PR #1368]: https://github.com/bareos/bareos/pull/1368
 [unreleased]: https://github.com/bareos/bareos/tree/master